TOT: Fixes glitch with inputting 8+ name
